+++ test/lib/Elastica/Test/Query/FunctionScoreTest.php
@@ -28,7 +28,7 @@ class FunctionScoreTest extends BaseTest
         $type->addDocuments([
             new Document(1, [
                 'name' => "Mr. Frostie's",
-                'location' => ['lat' => 32.799605, 'lon' => -117.243027],
+                'location' => [['lat' => 32.799605, 'lon' => -117.243027], ['lat' => 32.792744, 'lon' => -117.2387341]],
                 'price' => 4.5,
                 'popularity' => null,
             ]),
@@ -349,9 +349,21 @@ class FunctionScoreTest extends BaseTest
             $locationScale,
             null,
             null,
-            .5
+            .5,
+            null,
+            FunctionScore::MULTI_VALUE_MODE_AVG
         );
-        $query->addDecayFunction(FunctionScore::DECAY_GAUSS, 'price', $priceOrigin, $priceScale, null, null, 2);
+        $query->addDecayFunction(
+            FunctionScore::DECAY_GAUSS,
+            'price',
+            $priceOrigin,
+            $priceScale,
+            null,
+            null,
+            2,
+            null,
+            FunctionScore::MULTI_VALUE_MODE_MAX
+        );
         $expected = [
             'function_score' => [
                 'query' => $childQuery->toArray(),
@@ -362,6 +374,7 @@ class FunctionScoreTest extends BaseTest
                                 'origin' => $this->locationOrigin,
                                 'scale' => $locationScale,
                             ],
+                            'multi_value_mode'=>FunctionScore::MULTI_VALUE_MODE_AVG,
                         ],
                         'weight' => .5,
                     ],
@@ -371,6 +384,7 @@ class FunctionScoreTest extends BaseTest
                                 'origin' => $priceOrigin,
                                 'scale' => $priceScale,
                             ],
+                            'multi_value_mode'=>FunctionScore::MULTI_VALUE_MODE_MAX
                         ],
                         'weight' => 2,
                     ],
@@ -394,6 +408,31 @@ class FunctionScoreTest extends BaseTest
         // the document with the closest location and lowest price should be scored highest
         $result0 = $results[0]->getData();
         $this->assertEquals("Mr. Frostie's", $result0['name']);
+    }
+
+    /**
+     * @group functional
+     */
+    public function testGaussMultiValue()
+    {
+        $query = new FunctionScore();
+        $query->addDecayFunction(
+            FunctionScore::DECAY_GAUSS,
+            'location',
+            $this->locationOrigin,
+            '4mi',
+            null,
+            null,
+            null,
+            null,
+            FunctionScore::MULTI_VALUE_MODE_SUM
+        );
+        $response = $this->_getIndexForTest()->search($query);
+        $results = $response->getResults();
+
+        // the document with the sum of distances should be scored highest
+        $result0 = $results[0]->getData();
+        $this->assertEquals("Miller's Field", $result0['name']);
     }
 
     /**